refactor(test): extract story-a acceptance helpers to shared bdd/support/helpers.ts
Lets both the original acceptance test and new BDD step definitions reuse the same workspace/artifact setup. No behavior change to existing tests. Production path remains unmocked (ERR-025).
